js/xpconnect/tests/mochitest/test_bug1094930.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1094930
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1094930</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <iframe id="ifr"></iframe>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1094930">Mozilla Bug 1094930</a>
<p id="display"></p>
<script type="application/javascript">
  SimpleTest.waitForExplicitFinish();
  class XFoo extends frames[0].HTMLElement {
    connectedCallback() {
      ok(true, "connectedCallback was called");
      SimpleTest.finish();
    }
  };

  customElements.define.call(frames[0].customElements, "x-foo", XFoo);
  frames[0].document.firstChild.appendChild(new XFoo());
</script>
</body>
</html>
